Add Patchwork UTF-8 support
This routine will help ensure that the core code is dealing with a consistent encoding (UTF-8) to prevent corruption. Also provides fallback rountines for the basic mbstring and iconv functions if those extentions are not loaded. This change gives us "Unicode all the way" from input to database to output.

Composer update is required for this change.

Also changed in composer.json:
- doctrine/dbal to new version 2.4.1 minor updates
- symfony/yaml locked to 2.3.5 as issues encountered in 2.3.6