Recent change to security class resulted in two issues
First, the system should exit if an attempt to contaminate globals is detected, and that was removed. That functionality was re-added to the Xoops class that calls checkSuperglobals()

Second, the check for any string starting with 'xoops' has problems with things like 'XOOPS_TOKEN_REQUEST' which are required for normal use.

While looking at these issue, it seems the checks were case sensitive (based on array keys,) while PHP variables name are case insensitive. Fixed.