Making counter persistent, to avoid collisions
In a case where you are layout out radio buttons in a "tray" and so creating a series of radio button elements with the same name, they would previously all get the $id_ele of 1, and so their labels and ids would all be screwed up.  This change gives them unique ids, and labels, since the counter persists across all radio buttons in a given page load.