Remove XoopsPreload and Xoops::preload() references
Replacing with Xoops\Core\Events and Xoops::events()

This is prep work for making the event system restartable inside a PHP
session. An example of the issues is the install process, where several
modules are installed in a single session. The preload events are not
passed to the newly installed modules, since the list is established
when Core\Events is first instantiated, and the singleton pattern
effectively locks it as is.