Calm risky test errors in Assert
When an Xmf\Assert method is successful, there is no return to
test, but if it fails there is an exception to catch. Successful
assertions therefor result in an "Risky" error:
"This test did not perform any assertions"

Dummy assertions needed to show code is covered, as ignoring the
risky message marks it as having no coverage.

Mr. PHPUnit thinks this is OK :(